+++ bank-feeds/src/codatbankfeeds/models/operations/get_configuration.py
@@ -19,10 +19,10 @@ class GetConfigurationRequest:
 class GetConfigurationResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
-    raw_response: requests_http.Response = dataclasses.field()
-    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     configuration: Optional[shared_configuration.Configuration] = dataclasses.field(default=None)
     r"""Success"""